opencv/requirements.txt
opencv-python>=4.5
numpy>=1.20
flask
picamera[array]
toml
pyserial
# imutils
